fix: Only show activation panel when needed
