+++ frontend/src/layout/navigation-3000/sidepanel/panels/SidePanelSupport.tsx
@@ -1,12 +1,13 @@
 import { useActions, useValues } from 'kea'
 import { SupportForm, SupportFormButtons } from 'lib/components/Support/SupportForm'
-import { SidePanelTab, sidePanelLogic } from '../sidePanelLogic'
 import { supportLogic } from 'lib/components/Support/supportLogic'
 import { useEffect } from 'react'
+import { sidePanelStateLogic } from '../sidePanelStateLogic'
+import { SidePanelTab } from '~/types'
 
 export const SidePanelSupport = (): JSX.Element => {
-    const { closeSidePanel } = useActions(sidePanelLogic)
-    const { selectedTab } = useValues(sidePanelLogic)
+    const { closeSidePanel } = useActions(sidePanelStateLogic)
+    const { selectedTab } = useValues(sidePanelStateLogic)
 
     const theLogic = supportLogic({ onClose: () => closeSidePanel(SidePanelTab.Feedback) })
     const { title } = useValues(theLogic)
